+++ demos/slack_bot/slack_or_with_tool_calling.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 from slack_bolt import App
 from slack_bolt.adapter.socket_mode import SocketModeHandler
 
-from demos.components.open_router_client import OpenRouterClient
+from demos.components.open_router.open_router_client import OpenRouterClient
 from demos.tool_calling.tool_descriptors import (tools_rag_descriptor,
                                                  tools_search_descriptor,
                                                  tools_get_website_contents)
